Feat/refactoring 권지민

+++ src/components/ui/molecules/feed-card/FeedCardContainer.jsx
@@ -1,3 +1,4 @@
+import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
 import styled from 'styled-components';
@@ -10,6 +11,7 @@ import Toast from '@components/ui/atoms/toast/Toast';
 import { deleteSubject } from '@api/subjects/deleteSubject';
 import { useAsync } from '@hooks/useAsync';
 import { useCheckAnswerPath } from '@hooks/useCheckAnswerPath';
+import { useCheckAuth } from '@hooks/useCheckAuth';
 import { useCloseModal } from '@hooks/useCloseModal';
 import { useConfirmAlert } from '@hooks/useConfirmAlert';
 
@@ -25,18 +27,29 @@ const FeedCardContainer = ({
   setRequestType,
 }) => {
   const { isModalOpen, toggleModal } = useCloseModal();
-  const { isDeleteOpen, toggleDeleteCard } = useCloseModal();
+  // const [deleteAllSuccess, setDeleteAllSuccess] = useState(false);
   const { showConfirm, ConfirmAlertComponent } = useConfirmAlert();
   const { setAsyncFunction } = useAsync(deleteSubject);
   const navigate = useNavigate();
   const { path } = useCheckAnswerPath();
+  const { isUser } = useCheckAuth();
 
   const handleDeleteAllClick = async () => {
-    await setAsyncFunction(userId);
-    setRequestType('deleteAll');
-
-    localStorage.removeItem('userId', null);
-    navigate('/');
+    try {
+      if (isUser) {
+        await setAsyncFunction(userId);
+        setRequestType('deleteAll');
+        localStorage.removeItem('userId', null);
+        navigate('/', {
+          state: { deleteAllSuccess: true },
+        });
+      } else {
+        alert('삭제 권한이 없습니다.');
+        navigate('/list');
+      }
+    } catch (error) {
+      console.error('계정 삭제에 실패하였습니다.');
+    }
   };
 
   const handleCancelDeleteAllClick = () => {
@@ -86,7 +99,6 @@ const FeedCardContainer = ({
           content={`정말로 마음을 닫으시겠어요? \n 삭제된 계정 정보와 피드는 복구할 수 없습니다.`}
         />
         {isModalOpen && <Toast closeModal={toggleModal}>취소 되었습니다.</Toast>}
-        {isDeleteOpen && <Toast closeModal={toggleDeleteCard}>삭제 되었습니다.</Toast>}
       </PortalContainer>
     </>
   );
